Cross-reference problems fixed.
