+++ src/main/resources/application.properties
@@ -17,7 +17,7 @@ artie.chatgpt.api.maxtokens=500
 artie.chatgpt.api.url =https://api.openai.com/v1/engines/davinci/completions
 artie.webservices.login.url=http://${API_GATEWAY_HOST}:${API_GATEWAY_PORT}/users/login
 artie.webservices.student.updateCompetence.url=http://${API_GATEWAY_HOST}:${API_GATEWAY_PORT}/students/updateStudentCompetence
-artie.webservices.help.url=http://localhost:5000/help/help-model/api/v1/help-model
+artie.webservices.help.url=http://${API_GATEWAY_HOST}:${API_GATEWAY_PORT}/help/help-model/api/v1/help-model
 artie.webservices.prolog.query.url = http://${API_GATEWAY_HOST}:${API_GATEWAY_PORT}/prolog/pedagogicalprogram/query
 artie.webservices.emotional.url=http://${API_GATEWAY_HOST}:${API_GATEWAY_PORT}/emotional/emotional-model
 artie.webservices.emotional.queue=emotionalStateRequests
